split annotate and evaluate to separate commands
